[chore]: enforced referential integrity on database

+++ db/migrate/20200513183447_create_constructor_results.rb
@@ -1,8 +1,8 @@
 class CreateConstructorResults < ActiveRecord::Migration[6.0]
   def change
     create_table :constructor_results do |t|
-      t.integer :race_id, null: false
-      t.integer :constructor_id, null: false
+      t.bigint :race_id, null: false
+      t.bigint :constructor_id, null: false
       t.float :points
       t.string :status
