+++ db/migrate/20200513190123_create_constructor_standings.rb
@@ -1,8 +1,8 @@
 class CreateConstructorStandings < ActiveRecord::Migration[6.0]
   def change
     create_table :constructor_standings do |t|
-      t.integer :race_id, null: false
-      t.integer :constructor_id, null: false
+      t.bigint :race_id, null: false
+      t.bigint :constructor_id, null: false
       t.float :points, null: false
       t.integer :position
       t.string :position_text
