+++ db/migrate/20200513194548_create_driver_standings.rb
@@ -1,8 +1,8 @@
 class CreateDriverStandings < ActiveRecord::Migration[6.0]
   def change
     create_table :driver_standings do |t|
-      t.integer :race_id, null: false
-      t.integer :driver_id, null: false
+      t.bigint :race_id, null: false
+      t.bigint :driver_id, null: false
       t.float :points, null: false
       t.integer :position
       t.string :position_text
