+++ db/migrate/20200516100824_create_lap_times.rb
@@ -1,8 +1,8 @@
 class CreateLapTimes < ActiveRecord::Migration[6.0]
   def change
     create_table :lap_times do |t|
-      t.integer :race_id, null: false
-      t.integer :driver_id, null: false
+      t.bigint :race_id, null: false
+      t.bigint :driver_id, null: false
       t.integer :lap, null: false
       t.integer :position
       t.string :time
