+++ db/migrate/20200516102049_create_pit_stops.rb
@@ -1,8 +1,8 @@
 class CreatePitStops < ActiveRecord::Migration[6.0]
   def change
     create_table :pit_stops do |t|
-      t.integer :race_id, null: false
-      t.integer :driver_id, null: false
+      t.bigint :race_id, null: false
+      t.bigint :driver_id, null: false
       t.integer :stop, null: false
       t.integer :lap, null: false
       t.string :time, null: false
