+++ db/migrate/20200516103642_create_qualifyings.rb
@@ -1,9 +1,9 @@
 class CreateQualifyings < ActiveRecord::Migration[6.0]
   def change
     create_table :qualifyings do |t|
-      t.integer :race_id, null: false
-      t.integer :driver_id, null: false
-      t.integer :constructor_id, null: false
+      t.bigint :race_id, null: false
+      t.bigint :driver_id, null: false
+      t.bigint :constructor_id, null: false
       t.integer :number, null: false
       t.integer :position
       t.string :q1
